feat: add Operations Planner create mission card to /ops page
- Remove Mission nav tab (Target icon, /mission prefix, MISSION_PREFIXES)
- Delete /mission page routes (src/app/mission/ removed entirely)
- API routes at /api/mission/ kept untouched
- Add GET /api/mission/active — returns most recent mission with entries,
  stages, and constraints included
- CreateMissionCard component: title input + duration presets (30/75/90d)
  + custom input. Shows completed state when mission exists.
- OperationsPlanner wrapper: fetches active mission, renders
  CreateMissionCard above existing DailyDashboard
- /ops page updated to render OperationsPlanner instead of DailyDashboard
- Existing /ops daily dashboard functionality fully preserved
